remerged some files in order to reduce number of additional files
