+++ src/CMakeLists.txt
@@ -207,21 +207,17 @@ set(common_SRCS
 	object_properties.cpp
 	lua_types.cpp
 	lua_common.cpp
-	lua_enum.cpp
 	lua_content.cpp
 	lua_craft.cpp
 	lua_node.cpp
 	lua_item.cpp
-	lua_pseudorandom.cpp
 	lua_environment.cpp
 	lua_nodetimer.cpp
 	lua_perlin.cpp
 	lua_entity.cpp
 	lua_object.cpp
 	lua_nodemeta.cpp
-	lua_itemstack.cpp
 	lua_inventory.cpp
-	lua_detached_inv.cpp
 	scriptapi.cpp
 	script.cpp
 	log.cpp
