--- src/lua_detached_inv.h
@@ -1,74 +1,0 @@
-/*
-Minetest-c55
-Copyright (C) 2013 celeron55, Perttu Ahola <celeron55@gmail.com>
-
-This program is free software; you can redistribute it and/or modify
-it under the terms of the GNU Lesser General Public License as published by
-the Free Software Foundation; either version 2.1 of the License, or
-(at your option) any later version.
-
-This program is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU Lesser General Public License for more details.
-
-You should have received a copy of the GNU Lesser General Public License along
-with this program; if not, write to the Free Software Foundation, Inc.,
-51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
-*/
-
-#ifndef LUA_DETACHED_INV_H_
-#define LUA_DETACHED_INV_H_
-
-#include <iostream>
-
-extern "C" {
-#include <lua.h>
-}
-
-#include "serverobject.h"
-#include "inventory.h"
-
-/*****************************************************************************/
-/* Minetest interface                                                        */
-/*****************************************************************************/
-/* Detached inventory callbacks */
-// Return number of accepted items to be moved
-int scriptapi_detached_inventory_allow_move(lua_State *L,
-		const std::string &name,
-		const std::string &from_list, int from_index,
-		const std::string &to_list, int to_index,
-		int count, ServerActiveObject *player);
-// Return number of accepted items to be put
-int scriptapi_detached_inventory_allow_put(lua_State *L,
-		const std::string &name,
-		const std::string &listname, int index, ItemStack &stack,
-		ServerActiveObject *player);
-// Return number of accepted items to be taken
-int scriptapi_detached_inventory_allow_take(lua_State *L,
-		const std::string &name,
-		const std::string &listname, int index, ItemStack &stack,
-		ServerActiveObject *player);
-// Report moved items
-void scriptapi_detached_inventory_on_move(lua_State *L,
-		const std::string &name,
-		const std::string &from_list, int from_index,
-		const std::string &to_list, int to_index,
-		int count, ServerActiveObject *player);
-// Report put items
-void scriptapi_detached_inventory_on_put(lua_State *L,
-		const std::string &name,
-		const std::string &listname, int index, ItemStack &stack,
-		ServerActiveObject *player);
-// Report taken items
-void scriptapi_detached_inventory_on_take(lua_State *L,
-		const std::string &name,
-		const std::string &listname, int index, ItemStack &stack,
-		ServerActiveObject *player);
-
-/*****************************************************************************/
-/* Mod API                                                                   */
-/*****************************************************************************/
-int l_create_detached_inventory_raw(lua_State *L);
-
-#endif /* LUA_DETACHED_INV_H_ */
