--- src/lua_enum.cpp
@@ -1,156 +1,0 @@
-/*
-Minetest-c55
-Copyright (C) 2013 celeron55, Perttu Ahola <celeron55@gmail.com>
-
-This program is free software; you can redistribute it and/or modify
-it under the terms of the GNU Lesser General Public License as published by
-the Free Software Foundation; either version 2.1 of the License, or
-(at your option) any later version.
-
-This program is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU Lesser General Public License for more details.
-
-You should have received a copy of the GNU Lesser General Public License along
-with this program; if not, write to the Free Software Foundation, Inc.,
-51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
-*/
-
-#include "lua_enum.h"
-
-#include "itemdef.h"
-#include "nodedef.h"
-#include "craftdef.h"
-#include "biome.h"
-
-
-#include "lua_types.h"
-
-/*
-	EnumString definitions
-*/
-
-struct EnumString es_ItemType[] =
-{
-	{ITEM_NONE, "none"},
-	{ITEM_NODE, "node"},
-	{ITEM_CRAFT, "craft"},
-	{ITEM_TOOL, "tool"},
-	{0, NULL},
-};
-
-struct EnumString es_DrawType[] =
-{
-	{NDT_NORMAL, "normal"},
-	{NDT_AIRLIKE, "airlike"},
-	{NDT_LIQUID, "liquid"},
-	{NDT_FLOWINGLIQUID, "flowingliquid"},
-	{NDT_GLASSLIKE, "glasslike"},
-	{NDT_ALLFACES, "allfaces"},
-	{NDT_ALLFACES_OPTIONAL, "allfaces_optional"},
-	{NDT_TORCHLIKE, "torchlike"},
-	{NDT_SIGNLIKE, "signlike"},
-	{NDT_PLANTLIKE, "plantlike"},
-	{NDT_FENCELIKE, "fencelike"},
-	{NDT_RAILLIKE, "raillike"},
-	{NDT_NODEBOX, "nodebox"},
-	{0, NULL},
-};
-
-struct EnumString es_ContentParamType[] =
-{
-	{CPT_NONE, "none"},
-	{CPT_LIGHT, "light"},
-	{0, NULL},
-};
-
-struct EnumString es_ContentParamType2[] =
-{
-	{CPT2_NONE, "none"},
-	{CPT2_FULL, "full"},
-	{CPT2_FLOWINGLIQUID, "flowingliquid"},
-	{CPT2_FACEDIR, "facedir"},
-	{CPT2_WALLMOUNTED, "wallmounted"},
-	{0, NULL},
-};
-
-struct EnumString es_LiquidType[] =
-{
-	{LIQUID_NONE, "none"},
-	{LIQUID_FLOWING, "flowing"},
-	{LIQUID_SOURCE, "source"},
-	{0, NULL},
-};
-
-struct EnumString es_NodeBoxType[] =
-{
-	{NODEBOX_REGULAR, "regular"},
-	{NODEBOX_FIXED, "fixed"},
-	{NODEBOX_WALLMOUNTED, "wallmounted"},
-	{0, NULL},
-};
-
-struct EnumString es_CraftMethod[] =
-{
-	{CRAFT_METHOD_NORMAL, "normal"},
-	{CRAFT_METHOD_COOKING, "cooking"},
-	{CRAFT_METHOD_FUEL, "fuel"},
-	{0, NULL},
-};
-
-struct EnumString es_TileAnimationType[] =
-{
-	{TAT_NONE, "none"},
-	{TAT_VERTICAL_FRAMES, "vertical_frames"},
-	{0, NULL},
-};
-
-struct EnumString es_BiomeTerrainType[] =
-{
-	{BIOME_TERRAIN_NORMAL, "normal"},
-	{BIOME_TERRAIN_LIQUID, "liquid"},
-	{BIOME_TERRAIN_NETHER, "nether"},
-	{BIOME_TERRAIN_AETHER, "aether"},
-	{BIOME_TERRAIN_FLAT,   "flat"},
-	{0, NULL},
-};
-
-bool string_to_enum(const EnumString *spec, int &result,
-		const std::string &str)
-{
-	const EnumString *esp = spec;
-	while(esp->str){
-		if(str == std::string(esp->str)){
-			result = esp->num;
-			return true;
-		}
-		esp++;
-	}
-	return false;
-}
-
-/*bool enum_to_string(const EnumString *spec, std::string &result,
-		int num)
-{
-	const EnumString *esp = spec;
-	while(esp){
-		if(num == esp->num){
-			result = esp->str;
-			return true;
-		}
-		esp++;
-	}
-	return false;
-}*/
-
-int getenumfield(lua_State *L, int table,
-		const char *fieldname, const EnumString *spec, int default_)
-{
-	int result = default_;
-	string_to_enum(spec, result,
-			getstringfield_default(L, table, fieldname, ""));
-	return result;
-}
-
-
