+++ src/lua_environment.cpp
@@ -33,7 +33,6 @@ with this program; if not, write to the Free Software Foundation, Inc.,
 #include "lua_nodemeta.h"
 #include "lua_nodetimer.h"
 #include "lua_object.h"
-#include "lua_itemstack.h"
 #include "lua_common.h"
 #include "lua_item.h"
 #include "lua_node.h"
